requirements.txt
crewai
crewai[tools]
gradio
langchain
matplotlib
python-dotenv